- fixed the stormlib 32/64 bit issue - changed timer and general delta arguments from float seconds to int millis -> This changed the chaos that was the timer code into a pleasant and consistent API - made more use of object pools - fixed regeneration, as well as a multitude of timer- and timer-related- bugs / some minor pet bugs

+++ Core/WCell.Core/Timers/IUpdatable.cs
@@ -13,8 +13,8 @@ namespace WCell.Core.Timers
 		/// <summary>
 		/// Updates the object.
 		/// </summary>
-		/// <param name="dt">the time since the last update in seconds</param>
-		void Update(float dt);
+		/// <param name="dt">the time since the last update in millis</param>
+		void Update(int dt);
 	}
 
 	/// <summary>
@@ -44,7 +44,7 @@ namespace WCell.Core.Timers
 
 		#region IUpdatable Members
 
-		public void Update(float dt)
+		public void Update(int dt)
 		{
 			Callback();
 		}
